Polish: extract interface and add fix imageBg display error
