+++ pages/article/[...slug].vue
@@ -10,20 +10,7 @@
 </template>
 
 <script setup lang="ts">
-import type { ParsedContent } from '@nuxt/content/dist/runtime/types'
-
-interface articleInfo extends ParsedContent {
-  title: string
-  date: string
-  description: string
-  image: {
-    src: string
-    alt: string
-    url: string
-    photographer: string
-  }
-  tags: string[]
-}
+import { articleInfo } from '~~/utils/interfaces'
 
 const route = useRoute()
 const data = await queryContent<articleInfo>(route.path).findOne()
