+++ pages/article/index.vue
@@ -8,10 +8,8 @@
     <!-- The first article -->
     <NuxtLink v-for="(article, idx) in data?.slice(0, 1)" :key="idx" :to="article._path">
       <div class="group flex flex-row gap-10 my-10 <sm:(flex-col gap-2 my-5)">
-        <div
-          class="sm:(flex-1 h-250px) <sm:(h-250px) rounded-1rem overflow-hidden transition-all group-hover:(rounded-2rem)">
-          <ImageBg :src="article.image.src" :hash="article.image.hash" />
-        </div>
+        <ImageBg class="sm:(flex-1 h-250px) <sm:(h-250px) rounded-1rem transition-all group-hover:(rounded-2rem)"
+          :src="article.image.src" :hash="article.image.hash" />
         <div class="flex-1 flex flex-col transition-all">
           <p class="text-4xl <md:(text-2xl) <sm:(text-xl)">{{ article.title }}</p>
           <p class="my-0 text-gray-700 text-md <sm:(text-base font-thin)">{{ article.description }}</p>
@@ -34,5 +32,6 @@
 </template>
 
 <script setup lang='ts'>
-const { data } = await useAsyncData('article', () => queryContent('/article').sort({ date: -1 }).find())
+import { articleInfo } from "~~/utils/interfaces"
+const { data } = await useAsyncData('article', () => queryContent<articleInfo>('/article').sort({ date: -1 }).find())
 </script>
